build(docs): remove version trigger, use env var
Previously a push with a version tag would trigger the workflow twice.
The `--ignore-remote-status` flag was added to the mike commands to not
fail on concurrent runs of the workflow (the second run has a diverging
git history after the first one pushes).
This could still happen if there are multiple pushes before the workflow
finishes.
If a new docs release is pushed manually before the CI finishes it might
be overwritten now, whereas previously the CI would fail to push.